requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
fastapi==0.116.1
Flask==3.1.2
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.2.0b1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pandas-stubs==2.3.2.250827
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyodbc==5.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
# uvloop is not supported on Windows; omit it to avoid install failure
# uvloop==0.21.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xlrd==2.0.2
pdf2image==1.17.0
python-docx==1.1.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
